componentes a tsx
los componentes se cambiaron con la extension tsx .

instaacion y migracion de tsx: cd frontend
yarn add -D typescript @types/react @types/react-dom @types/node

en la terminal de nuestro frontend
yarn tsc --init

todas  las instalacion para migrar a typescript
