+++ frontend/package.json
@@ -6,14 +6,20 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
+    "@types/axios": "^0.14.4",
+    "@types/bootstrap": "^5.2.10",
     "axios": "^1.7.9",
+    "boostrap": "^2.0.0",
     "bootstrap": "^5.3.3",
+    "i": "^0.3.7",
     "react": "^19.0.0",
+    "react-bootstrap": "^2.10.9",
     "react-dom": "^19.0.0",
     "react-icons": "^5.4.0",
     "react-router-dom": "^7.1.3",
     "react-scripts": "5.0.1",
     "recharts": "^2.15.1",
+    "require": "^2.4.20",
     "tailwindcss": "^3.4.17",
     "web-vitals": "^2.1.0"
   },
@@ -42,5 +48,11 @@
     ]
   },
   "main": "app.js",
-  "license": "MIT"
+  "license": "MIT",
+  "devDependencies": {
+    "@types/node": "^22.13.13",
+    "@types/react": "^19.0.12",
+    "@types/react-dom": "^19.0.4",
+    "typescript": "^5.8.2"
+  }
 }
