+++ frontend/src/components/ajustes.tsx
