+++ frontend/src/components/comparacion.tsx
