+++ frontend/src/components/escala.tsx
