+++ frontend/src/components/infocompar.tsx
