+++ frontend/src/components/infonivelsatis.tsx
@@ -11,7 +11,7 @@ import {
   Tooltip,
 } from "recharts";
 
-const Infonivelsatis = () => {
+const Infonivelsatis:React.FC = () => {
   return (
     <Home>
       <div className="container">
@@ -45,7 +45,7 @@ const Infonivelsatis = () => {
             <div className="col-sm ">
               <div className="d-flex flex-row">
                 <div className="p-2">
-                  <Cuadrado color="green" width="30px" height="30px" />
+                  <Cuadrado color="green" width="30px" height="30px" letra="A" />
                 </div>
                 <div className="p-2">Opinion favorables (4 a 5)</div>
               </div>
@@ -53,7 +53,7 @@ const Infonivelsatis = () => {
             <div className="col-sm">
               <div className="d-flex flex-row">
                 <div className="p-2">
-                  <Cuadrado color="green" width="30px" height="30px" />{" "}
+                  <Cuadrado color="green" width="30px" height="30px" letra="A" />{" "}
                   {/* Cuadrado verde */}
                 </div>
                 <div className="p-2">Opinion favorables (4 a 5)</div>
@@ -70,7 +70,7 @@ const Infonivelsatis = () => {
             <div className="col-sm">
               <div className="d-flex flex-row">
                 <div className="p-2">
-                  <Cuadrado color="yellow" width="30px" height="30px" />{" "}
+                  <Cuadrado color="yellow" width="30px" height="30px" letra="A" />{" "}
                   {/* Cuadrado amarillo */}
                 </div>
                 <div className="p-2">Opinion indecisa (2 a 3.9)</div>
@@ -79,7 +79,7 @@ const Infonivelsatis = () => {
             <div className="col-sm">
               <div className="d-flex flex-row">
                 <div className="p-2">
-                  <Cuadrado color="yellow" width="30px" height="30px" />{" "}
+                  <Cuadrado color="yellow" width="30px" height="30px" letra="A" />{" "}
                   {/* Cuadrado amarillo */}
                 </div>
                 <div className="p-2">Opinion indecisa (2 a 3.9)</div>
@@ -96,7 +96,7 @@ const Infonivelsatis = () => {
             <div className="col-sm">
               <div className="d-flex flex-row">
                 <div className="p-2">
-                  <Cuadrado color="red" width="30px" height="30px" />{" "}
+                  <Cuadrado color="red" width="30px" height="30px" letra="A" />{" "}
                   {/* Cuadrado rosa */}
                 </div>
                 <div className="p-2">Opinion Desafavorables (1 a 1.9)</div>
@@ -106,7 +106,7 @@ const Infonivelsatis = () => {
             <div className="col-sm">
               <div className="d-flex flex-row">
                 <div className="p-2">
-                  <Cuadrado color="red" width="30px" height="30px" />{" "}
+                  <Cuadrado color="red" width="30px" height="30px" letra="A" />{" "}
                   {/* Cuadrado rosa */}
                 </div>
                 <div className="p-2">Opinion Desafavorables (1 a 1.9)</div>
