+++ frontend/src/components/listapuestos.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const listapuestos = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default listapuestos
