+++ frontend/src/components/nav-sidebar/Navtop.tsx
@@ -3,50 +3,48 @@ import { IoIosNotifications } from "react-icons/io";
 import { FaRegUserCircle } from "react-icons/fa";
 import { GiHamburgerMenu } from "react-icons/gi";
 import { Link } from "react-router-dom";
-import bootstrap from "bootstrap/dist/js/bootstrap.bundle.min"; // Asegúrate de que esto está importado
 
-const Navtop = ({ toggleSidebar }) => {
+interface NavtopProps {
+  toggleSidebar: () => void; // toggleSidebar es una función que no retorna nada
+}
 
-  useEffect(() => {
-    // Inicializar los popovers
-    const popoverTriggerList = [].slice.call(document.querySelectorAll('[data-bs-toggle="popover"]'));
-    const popovers = popoverTriggerList.map((popoverTriggerEl) => {
-      return new bootstrap.Popover(popoverTriggerEl);
-    });
-
-    // Limpiar los popovers al desmontar el componente o cuando cambian las rutas
-    return () => {
-      popovers.forEach((popover) => {
-        popover.dispose();  // Destruye todos los popovers
-      });
-    };
-  }, []); // Se ejecuta solo una vez después de que el componente se haya montado
-
+const Navtop: React.FC<NavtopProps> = ({ toggleSidebar }) => {
+ 
   return (
     <nav className="navbar border border-secondary navbar-dark text-bg-light position-sticky top-0 w-100">
       <div className="d-flex justify-content-between align-items-center w-100 pe-4">
         <div className="w-75">
-          <button className="btn btn-lg btn-light ms-2" onClick={toggleSidebar}>
+          <button
+            className="btn btn-lg btn-light ms-2"
+            onClick={toggleSidebar}
+            aria-label="Toggle Sidebar"
+          >
             <GiHamburgerMenu />
           </button>
           <Link to="/periodo">
-            <img className="img-fluid" style={{ width: "250px" }} src="/workfeel-logo.png" alt="Workfeel Logo" />
+            <img
+              className="img-fluid"
+              style={{ width: "250px" }}
+              src="/workfeel-logo.png"
+              alt="Workfeel Logo"
+            />
           </Link>
         </div>
 
         {/* Ícono de notificación, Perfil y nombre */}
         <div className="d-flex align-items-center">
-          <button 
+          <button
             type="button"
             className="btn"
-            data-bs-toggle="popover" 
-            title="Notificaciones" 
+            data-bs-toggle="popover"
+            title="Notificaciones"
             data-bs-content="No tienes notificaciones por leer."
             style={{
               background: "transparent",
               border: "none",
               marginLeft: "10px",
             }}
+            aria-label="Notificaciones"
           >
             <IoIosNotifications style={{ fontSize: "30px" }} />
           </button>
@@ -64,6 +62,7 @@ const Navtop = ({ toggleSidebar }) => {
                 padding: "5px",
                 marginLeft: "10px",
               }}
+              aria-label="Opciones de usuario"
             >
               <FaRegUserCircle style={{ fontSize: "28px" }} />
             </button>
@@ -72,25 +71,25 @@ const Navtop = ({ toggleSidebar }) => {
               aria-labelledby="dropdownMenu2"
             >
               <li>
-                <button className="dropdown-item" type="button">
-                  <Link to="/Ajustes">Ajustes</Link>
-                </button>
+                <Link to="/Ajustes" className="dropdown-item">
+                  Ajustes
+                </Link>
               </li>
               <li>
-                <button className="dropdown-item" type="button">
-                  <Link to="/perfil">Perfil</Link>
-                </button>
+                <Link to="/perfil" className="dropdown-item">
+                  Perfil
+                </Link>
               </li>
               <li>
-                <button className="dropdown-item" type="button">
-                  <Link to="/consumo">Consumo</Link>
-                </button>
+                <Link to="/consumo" className="dropdown-item">
+                  Consumo
+                </Link>
               </li>
             </ul>
           </div>
           <span className="ms-2" style={{ fontSize: "16px" }}>
-                Crhistopher Gutierrez
-              </span>{" "}
+            Crhistopher Gutierrez
+          </span>
         </div>
       </div>
     </nav>
