+++ frontend/src/components/nivelsatisfaccion.tsx
