+++ frontend/src/components/periodo.tsx
@@ -7,95 +7,7 @@ import { MdEdit, MdDelete } from "react-icons/md";
 import { BsExclamationTriangle } from "react-icons/bs";
 
 const Periodo = () => {
-  // Estado para manejar los periodos
-  const [periodos, setPeriodos] = useState([
-    {
-      id: 1,
-      nombre: "Periodo 1",
-      estado: "Cerrado",
-      inicio: "2023-01-01",
-      cierre: "2023-01-31",
-    },
-    {
-      id: 2,
-      nombre: "Periodo 2",
-      estado: "Cerrado",
-      inicio: "2024-01-01",
-      cierre: "2024-01-31",
-    },
-    {
-      id: 3,
-      nombre: "Periodo 3",
-      estado: "Cerrado",
-      inicio: "2024-01-01",
-      cierre: "2024-01-31",
-    },
-    {
-      id: 4,
-      nombre: "Periodo 4",
-      estado: "Cerrado",
-      inicio: "2024-01-01",
-      cierre: "2024-01-31",
-    },
-  ]);
 
-  //estado para agregar nuevos periodos en la table
-  const [nuevoPeriodo, setNuevoPeriodo] = useState({
-    nombre: "",
-    estado: "Cerrado",
-    inicio: "",
-    cierre: "",
-  });
-
-  const handleNuevoPeriodoChange = (e) => {
-    const { name, value } = e.target;
-    setNuevoPeriodo((prevPeriodo) => ({
-      ...prevPeriodo,
-      [name]: value,
-    }));
-  };
-
-  const handleAgregarPeriodo = () => {
-    const nuevoId = periodos.length ? periodos[periodos.length - 1].id + 1 : 1;
-    setPeriodos([...periodos, { id: nuevoId, ...nuevoPeriodo }]);
-    setNuevoPeriodo({ nombre: "", estado: "Cerrado", inicio: "", cierre: "" });
-  };
-
-  // Función para agregar un nuevo periodo
-
-  // Estado para manejar la búsqueda
-  const [busqueda, setBusqueda] = useState("");
-
-  // Estado para manejar el id del periodo a eliminar
-  const [periodoBorrar, setPeriodoBorrar] = useState(null);
-
-  // Estado para manejar el periodo que se está editando
-  const [periodoEditar, setPeriodoEditar] = useState(null);
-
-  // Filtrar periodos según la búsqueda
-  const periodosFiltrados = periodos.filter((p) =>
-    p.nombre.toLowerCase().includes(busqueda.toLowerCase())
-  );
-
-  // Función para manejar la eliminación de un periodo
-  const handleBorrarPeriodo = (id) => {
-    setPeriodos(periodos.filter((periodo) => periodo.id !== id));
-  };
-
-  // Función para manejar la edición de un periodo
-  const handleEditarPeriodo = (periodo) => {
-    setPeriodoEditar(periodo);
-  };
-
-  // Función para guardar los cambios del periodo editado
-  const handleGuardarCambios = () => {
-    setPeriodos(
-      periodos.map((periodo) =>
-        periodo.id === periodoEditar.id ? periodoEditar : periodo
-      )
-    );
-    setPeriodoEditar(null);
-  };
 
   return (
     <Home>
@@ -155,8 +67,6 @@ const Periodo = () => {
                 className="form-control form-control-sm"
                 placeholder="Buscar..."
                 aria-label="Buscar"
-                value={busqueda}
-                onChange={(e) => setBusqueda(e.target.value)}
               />
             </div>
           </div>
@@ -182,47 +92,6 @@ const Periodo = () => {
               </tr>
             </thead>
             <tbody>
-              {periodosFiltrados.map((periodo) => (
-                <tr key={periodo.id}>
-                  <th scope="row">{periodo.nombre}</th>
-                  <td>{periodo.estado}</td>
-                  <td>{periodo.inicio}</td>
-                  <td>{periodo.cierre}</td>
-                  <td>
-                    <div className="btn-group">
-                      <button
-                        type="button"
-                        className="btn"
-                        data-bs-toggle="modal"
-                        data-bs-target="#editModal"
-                        onClick={() => handleEditarPeriodo(periodo)}
-                      >
-                        <MdEdit className="me-2" />
-                        Editar
-                      </button>
-                      <button
-                        type="button"
-                        className="btn"
-                        data-bs-toggle="modal"
-                        data-bs-target="#deleteModal"
-                        onClick={() => setPeriodoBorrar(periodo.id)}
-                      >
-                        <MdDelete className="me-2" />
-                        Eliminar
-                      </button>
-                      <button
-                        type="button"
-                        className="btn"
-                        data-bs-toggle="modal"
-                        data-bs-target="#sharemodel"
-                      >
-                        <IoShareSocialOutline className="me-2" />
-                        Compartir
-                      </button>
-                    </div>
-                  </td>
-                </tr>
-              ))}
             </tbody>
           </table>
         </div>
@@ -231,7 +100,7 @@ const Periodo = () => {
         <div
           className="modal fade"
           id="editModal"
-          tabIndex="-1"
+          tabIndex={-1}
           aria-hidden="true"
         >
           <div className="modal-dialog">
@@ -255,13 +124,6 @@ const Periodo = () => {
                   id="edita"
                   placeholder="Periodo 1"
                   className="border rounded text-center w-75 h-100 h6"
-                  value={periodoEditar?.nombre || ""}
-                  onChange={(e) =>
-                    setPeriodoEditar({
-                      ...periodoEditar,
-                      nombre: e.target.value,
-                    })
-                  }
                 />
                 <div className="row">
                   <div className="col m-3 h6">
@@ -272,13 +134,7 @@ const Periodo = () => {
                       type="date"
                       id="editar"
                       className="border rounded text-center w-100 h6"
-                      value={periodoEditar?.inicio || ""}
-                      onChange={(e) =>
-                        setPeriodoEditar({
-                          ...periodoEditar,
-                          inicio: e.target.value,
-                        })
-                      }
+                      
                     />
                   </div>
                   <div className="col m-3 h6">
@@ -289,13 +145,7 @@ const Periodo = () => {
                       type="date"
                       id="editar"
                       className="border rounded text-center w-100 h6"
-                      value={periodoEditar?.cierre || ""}
-                      onChange={(e) =>
-                        setPeriodoEditar({
-                          ...periodoEditar,
-                          cierre: e.target.value,
-                        })
-                      }
+                    
                     />
                   </div>
                   <label htmlFor="estado " className="opacity-50 h6 m-1">
@@ -304,13 +154,7 @@ const Periodo = () => {
                   <select
                     id="estado"
                     className="border rounded text-center w-75 m-3 opacity-50 h6"
-                    value={periodoEditar?.estado || ""}
-                    onChange={(e) =>
-                      setPeriodoEditar({
-                        ...periodoEditar,
-                        estado: e.target.value,
-                      })
-                    }
+                    
                   >
                     <option value="Cerrado">Cerrado</option>
                     <option value="Activo">Activo</option>
@@ -324,7 +168,6 @@ const Periodo = () => {
                   type="button"
                   className="btn btn-primary"
                   data-bs-dismiss="modal"
-                  onClick={handleGuardarCambios}
                 >
                   Guardar cambios
                 </button>
@@ -344,7 +187,7 @@ const Periodo = () => {
         <div
           className="modal fade"
           id="deleteModal"
-          tabIndex="-1"
+          tabIndex={-1}
           aria-hidden="true"
         >
           <div className="modal-dialog">
@@ -381,7 +224,6 @@ const Periodo = () => {
                   type="button"
                   className="btn btn-danger"
                   data-bs-dismiss="modal"
-                  onClick={() => handleBorrarPeriodo(periodoBorrar)}
                 >
                   Eliminar
                 </button>
@@ -401,7 +243,7 @@ const Periodo = () => {
         <div
           className="modal fade"
           id="nuevoPeriodoModal"
-          tabIndex="-1"
+          tabIndex={-1}
           aria-hidden="true"
         >
           <div className="modal-dialog ">
@@ -429,8 +271,7 @@ const Periodo = () => {
                   id="nombre"
                   name="nombre"
                   className="border rounded text-center w-75 h-100 h6"
-                  value={nuevoPeriodo.nombre}
-                  onChange={handleNuevoPeriodoChange}
+                 
                 />
 
                 <div className="row">
@@ -443,8 +284,7 @@ const Periodo = () => {
                       id="inicio"
                       name="inicio"
                       className="border rounded text-center w-100 h6"
-                      value={nuevoPeriodo.inicio}
-                      onChange={handleNuevoPeriodoChange}
+                      
                     />
                   </div>
                   <div className="col m-3 h6">
@@ -456,8 +296,7 @@ const Periodo = () => {
                       id="cierre"
                       name="cierre"
                       className="border rounded text-center w-100 h6"
-                      value={nuevoPeriodo.cierre}
-                      onChange={handleNuevoPeriodoChange}
+                     
                     />
                   </div>
                   <label htmlFor="estado " className="opacity-50 h6 m-1">
@@ -467,8 +306,7 @@ const Periodo = () => {
                     id="estado"
                     name="estado"
                     className="border rounded text-center w-75 m-3 opacity-50 h6"
-                    value={nuevoPeriodo.estado}
-                    onChange={handleNuevoPeriodoChange}
+                    
                   >
                     <option value="Cerrado">Cerrado</option>
                     <option value="Activo">Activo</option>
@@ -481,7 +319,7 @@ const Periodo = () => {
                   type="button"
                   className="btn btn-primary"
                   data-bs-dismiss="modal"
-                  onClick={handleAgregarPeriodo}
+                  
                 >
                   Guardar cambios
                 </button>
