+++ frontend/src/components/planaccion.tsx
