+++ frontend/src/components/resultados.tsx
@@ -386,7 +386,7 @@ const Resultados = () => {
             <div className="col"></div>
           </div>
         </div>
-        <listapuestos />
+
       </div>
     </Home>
   );
