+++ frontend/src/components/tableinterpretacion.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Cuadrado from "./cuadrado";
 
-const Tableinterpretacion = () => {
+const Tableinterpretacion:React.FC = () => {
   return (
     <div className="container">
       <div className="border text-center m-3">
@@ -34,7 +34,8 @@ const Tableinterpretacion = () => {
           <div className="col-sm ">
             <div className="d-flex flex-row">
               <div className="p-2">
-                <Cuadrado color="green" width="30px" height="30px" />
+               
+                <Cuadrado color="green" width="30px" height="30px" letra="A" />
               </div>
               <div className="p-2">Opinion favorables (4 a 5)</div>
             </div>
@@ -42,7 +43,8 @@ const Tableinterpretacion = () => {
           <div className="col-sm">
             <div className="d-flex flex-row">
               <div className="p-2">
-                <Cuadrado color="green" width="30px" height="30px" />{" "}
+                
+                <Cuadrado color="green" width="30px" height="30px" letra="A"/>{" "}
                 {/* Cuadrado verde */}
               </div>
               <div className="p-2">Opinion favorables (4 a 5)</div>
@@ -59,7 +61,7 @@ const Tableinterpretacion = () => {
           <div className="col-sm">
             <div className="d-flex flex-row">
               <div className="p-2">
-                <Cuadrado color="yellow" width="30px" height="30px" />{" "}
+                <Cuadrado color="yellow" width="30px" height="30px" letra="A"/>{" "}
                 {/* Cuadrado amarillo */}
               </div>
               <div className="p-2">Opinion indecisa (2 a 3.9)</div>
@@ -68,7 +70,7 @@ const Tableinterpretacion = () => {
           <div className="col-sm">
             <div className="d-flex flex-row">
               <div className="p-2">
-                <Cuadrado color="yellow" width="30px" height="30px" />{" "}
+                <Cuadrado color="yellow" width="30px" height="30px" letra="A"/>{" "}
                 {/* Cuadrado amarillo */}
               </div>
               <div className="p-2">Opinion indecisa (2 a 3.9)</div>
@@ -85,7 +87,7 @@ const Tableinterpretacion = () => {
           <div className="col-sm">
             <div className="d-flex flex-row">
               <div className="p-2">
-                <Cuadrado color="red" width="30px" height="30px" />{" "}
+                <Cuadrado color="red" width="30px" height="30px" letra="A"/>{" "}
                 {/* Cuadrado rosa */}
               </div>
               <div className="p-2">Opinion Desafavorables (1 a 1.9)</div>
@@ -95,7 +97,7 @@ const Tableinterpretacion = () => {
           <div className="col-sm">
             <div className="d-flex flex-row">
               <div className="p-2">
-                <Cuadrado color="red" width="30px" height="30px" />{" "}
+                <Cuadrado color="red" width="30px" height="30px" letra="A"/>{" "}
                 {/* Cuadrado rosa */}
               </div>
               <div className="p-2">Opinion Desafavorables (1 a 1.9)</div>
