+++ frontend/src/components/triangulo.tsx
